+++ essentialism-auth/authModel.js
@@ -1,0 +1,10 @@
+const db = require('../data/dbConfig');
+
+function addUser(user) {
+    return db('users')
+        .insert(user, 'id')
+}
+
+module.exports = {
+    addUser
+}
