+++ essentialism/essentialismModel.js
@@ -1,10 +1,5 @@
 const db = require('../data/dbConfig');
 
-function addUser(user) {
-    return db('users')
-        .insert(user, 'id')
-}
-
 function findValues(id) {
     return db('values')
         .join('users', 'users.id', 'values.userId')
@@ -20,6 +15,5 @@ function findValues(id) {
 };
 
 module.exports = {
-    addUser,
     findValues
 }
