+++ essentialism/essentialismRouter.js
@@ -1,27 +1,8 @@
 const express = require('express');
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const secrets = require('../secrets');
 
 const Essentialism = require('./essentialismModel');
 
 const router = express.Router();
-
-router.post('/register', (req, res) => {
-    let user = req.body;
-    const hash = bcrypt.hashSync(user.password, 10);
-    user.password = hash;
-
-    Essentialism.addUser(user)
-        .then(info => {
-            res.status(201).json(info);
-        })
-        .catch(err => {
-            res.status(500).json({
-                error: 'Failed to create user'
-            })
-        })
-})
 
 router.get('/values/:id', (req, res) => {
     const {id} = req.params;
